gae/requirements.txt
Flask==0.12.2
google-api-python-client==1.6.5
google-auth==1.4.0
google-auth-httplib2==0.0.3
google-api-python-client==1.6.5
google-cloud-datastore==1.4.0
google-cloud-storage==1.7.0
google-cloud-bigquery==0.31.0
oauth2client==4.1.2
gunicorn==19.7.1
apache-beam==2.3.0
